Déplacement de datePublication et lienPublication de "declaration" vers "informationsComplementaires"
